backend/pyproject.toml
[project]
name = "apexfit-backend"
version = "0.1.0"
description = "ApexFit FastAPI backend for health & fitness tracking"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "firebase-admin>=6.6.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "langchain>=0.3.0",
    "langchain-google-vertexai>=2.0.0",
    "google-cloud-tasks>=2.16.0",
    "google-cloud-storage>=2.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"app/engines/config.py" = ["N815"]

[tool.mypy]
python_version = "3.11"
strict = true
